Added configuration for global user id thought the app.

+++ src/Services/Follows/InstaConnect.Follows.Business/Commands/Follows/DeleteFollow/DeleteFollowCommandHandler.cs
@@ -1,27 +1,23 @@
 ﻿using AutoMapper;
 using InstaConnect.Follows.Data.Abstractions.Repositories;
+using InstaConnect.Shared.Business.Abstractions;
+using InstaConnect.Shared.Business.Exceptions.Account;
 using InstaConnect.Shared.Business.Exceptions.Follow;
 using InstaConnect.Shared.Business.Messaging;
-using InstaConnect.Shared.Business.Models.Requests;
-using InstaConnect.Shared.Business.Models.Responses;
-using MassTransit;
 
 namespace InstaConnect.Follows.Business.Commands.Follows.DeleteFollow;
 
 public class DeleteFollowCommandHandler : ICommandHandler<DeleteFollowCommand>
 {
-    private readonly IMapper _mapper;
     private readonly IFollowRepository _followRepository;
-    private readonly IRequestClient<ValidateUserByIdRequest> _validateUserByIdRequestClient;
+    private readonly ICurrentUserContext _currentUserContext;
 
     public DeleteFollowCommandHandler(
-        IMapper mapper,
         IFollowRepository followRepository,
-        IRequestClient<ValidateUserByIdRequest> validateUserByIdRequestClient)
+        ICurrentUserContext currentUserContext)
     {
-        _mapper = mapper;
         _followRepository = followRepository;
-        _validateUserByIdRequestClient = validateUserByIdRequestClient;
+        _currentUserContext = currentUserContext;
     }
 
     public async Task Handle(DeleteFollowCommand request, CancellationToken cancellationToken)
@@ -33,8 +29,12 @@ public class DeleteFollowCommandHandler : ICommandHandler<DeleteFollowCommand>
             throw new FollowNotFoundException();
         }
 
-        var validateUserByIdRequest = _mapper.Map<ValidateUserByIdRequest>(existingFollow);
-        await _validateUserByIdRequestClient.GetResponse<ValidateUserByIdResponse>(validateUserByIdRequest, cancellationToken);
+        var currentUserDetails = _currentUserContext.GetCurrentUserDetails();
+
+        if (currentUserDetails.Id != existingFollow.FollowerId)
+        {
+            throw new AccountForbiddenException();
+        }
 
         await _followRepository.DeleteAsync(existingFollow, cancellationToken);
     }
